разбиение main на модули
